Rename @spectrum-i18n/message-formatter to just @spectrum-i18n/message
