+++ packages/@react-stately/color/package.json
@@ -24,7 +24,7 @@
     "@react-stately/utils": "^3.1.0",
     "@react-types/color": "3.0.0-alpha.1",
     "@react-types/numberfield": "3.0.0-alpha.1",
-    "@spectrum-i18n/message-formatter": "3.0.0-alpha.1",
+    "@spectrum-i18n/message": "3.0.0-alpha.1",
     "@spectrum-i18n/number": "3.0.0-alpha.1"
   },
   "peerDependencies": {
