+++ packages/@react-stately/color/src/Color.ts
@@ -14,7 +14,7 @@ import {clamp} from '@react-aria/utils';
 import {ColorChannel, ColorChannelRange, ColorFormat, Color as IColor} from '@react-types/color';
 // @ts-ignore
 import intlMessages from '../intl/*.json';
-import {MessageDictionary} from '@spectrum-i18n/message-formatter';
+import {MessageDictionary} from '@spectrum-i18n/message';
 import {NumberFormatter} from '@spectrum-i18n/number';
 
 const messages = new MessageDictionary(intlMessages);
