+++ packages/@spectrum-i18n/message/src/index.ts
@@ -10,4 +10,5 @@
  * governing permissions and limitations under the License.
  */
 
+export * from './MessageDictionary';
 export * from './MessageFormatter';
